+++ CHANGELOG.md
@@ -19,6 +19,10 @@ To install release candidates run `[sudo] gem install cocoapods --pre`
 
 ##### Bug Fixes
 
+* Fix adding developer library search paths during pod validation.  
+  [Nick Entin](https://github.com/NickEntin)
+  [#9736](https://github.com/CocoaPods/CocoaPods/pull/9736)
+
 * Fix an issue that caused multiple xcframework scripts to produce the same output files  
   [Eric Amorde](https://github.com/amorde)
   [#9670](https://github.com/CocoaPods/CocoaPods/issues/9670)